requirements.txt
# Core API
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9

# Data Processing
pandas==2.2.0
numpy==1.26.3
sodapy==2.2.0  # SF Open Data API client

# ML Models
scikit-learn==1.4.0
xgboost==2.0.3
prophet==1.1.5  # Time series forecasting
joblib==1.3.2

# Geospatial
geopandas==0.14.2
shapely==2.0.2
pyproj==3.6.1

# GCP
google-cloud-bigquery==3.17.1
google-cloud-storage==2.14.0
google-cloud-logging==3.9.0
google-auth==2.27.0

# Visualization & Data
plotly==5.18.0
folium==0.15.1

# Utilities
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
httpx==0.26.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Scheduling
apscheduler==3.10.4

# Date/Time
python-dateutil==2.8.2
pytz==2024.1
